Fix callback concurrency issues; ensure more methods will resolve callbacks; do a best-case handling for userCompletedAction; update readme
1. No more class-level callbackContext; individual callbackContexts are
passed between methods to ensure we don’t clobber the callbackContext
with two concurrent calls to execute()
2. logout, registerView, and userCompletedAction were missing callback
resolutions, so I added those. userCompletedAction’s logic is not
completely correct as it will always (early) resolve to “success” but
Branch.java in the Branch SDK does not expose a method that takes in a
listener for userCompletedAction that I saw.
3. Fixed CreditHistoryListener to reject a promise if there was a JSON
exception
4. Added properly evaluated callback to setIdentity
5. Updated Readme.md to match code for installation