Remove install referrer intent on Android
Per branch email: 

Dear Customer,

At the beginning of this month, we sent you an email with the subject “Please confirm your Play Store Referrer configuration”. We want to make a correction to this email.

Correction: if you chose to update to the newest Android SDK, then please be sure to delete the entry from AndroidManifest.xml related to the install_referrer intent broadcast. Failure to do so will result in warnings in your development console and crashes on new installs via the Play Store. Previously we referred to this step as recommended; however, it is required if you upgrade to the latest versions of the SDK (Branch Android: v4.3.2; TUNE Android: v6.1.2).

Best,

The Branch Team